# Optimize redundant shared pointer copying/moving during constant vector wrapping.