Summary:
before:
```
============================================================================
fbcode/velox/benchmarks/basic/Preproc.cpp     relative  time/iter   iters/s
============================================================================
ifFloor                                                     7.46ms    134.02
ifFloorWithSimpleEq                                         8.83ms    113.24
oneHot                                                      3.30ms    302.65
oneHotWithVectorArithmetic                                  3.27ms    306.23
allFused                                                    1.16ms    862.31
----------------------------------------------------------------------------
ifFloorWithNulls                                           24.25ms     41.23
ifFloorWithSimpleEqWithNulls                               20.01ms     49.97
oneHotWithNulls                                            14.40ms     69.42
oneHotWithVectorArithmeticWithNulls                        14.43ms     69.32
allFusedWithNulls                                           4.05ms    247.08

```

after:

```
============================================================================
fbcode/velox/benchmarks/basic/Preproc.cpp     relative  time/iter   iters/s
============================================================================
ifFloor                                                     6.93ms    144.37
ifFloorWithSimpleEq                                         8.27ms    120.90
oneHot                                                      3.28ms    305.13
oneHotWithVectorArithmetic                                  3.27ms    305.88
allFused                                                    1.20ms    835.53
----------------------------------------------------------------------------
ifFloorWithNulls                                           22.31ms     44.83
ifFloorWithSimpleEqWithNulls                               16.84ms     59.37
oneHotWithNulls                                            12.98ms     77.01
oneHotWithVectorArithmeticWithNulls                        12.58ms     79.52
allFusedWithNulls                                           3.71ms    269.58

```

Differential Revision: D41879701